cleaned up mock backend code. Better shutdown
